Fix 30+ porting issues across Java, AssemblyScript, and Go SDKs
- Java: fix TeaVM build (buildscript classpath, nested wasm DSL), add
  DurableEntryAggregator for tree-shaking, add getQueryState, add Saga
- AS: fix encodeUnsafe/type errors for AS 0.27.32, fix transform
  SUSPEND_SENTINEL handling and try/catch removal, add escapeJson,
  add Saga, fix scoped imports via asconfig.json paths
- Go: fix DurableCallJSON nil result panic, change Saga Compensate to
  return error, add DurableCallTypedWithOptions, add child workflow
  stubbing in TestEnv, fix Saga error wrapping
- Examples: fix as-workflow (imports, API names, --transform flag),
  update fooddash/travel Saga signatures
- Docs: add READMEs for Java and AS SDKs with known constraints

+++ packages/cleat-as/assembly/host-calls.ts
@@ -417,8 +417,8 @@ export class HostCalls {
         "scratch buffer overflow: no space remaining for '" + label + "'",
       );
     }
-    let byteLen: usize = String.UTF8.byteLength(s);
-    if (byteLen > remaining as usize) {
+    let byteLen: i32 = String.UTF8.byteLength(s) as i32;
+    if (byteLen > remaining) {
       throw new Error(
         "scratch buffer overflow: '" +
           label +
